ShellPkg/AcpiView: Detect duplicate fields and invalid references
Define an implement methods for validating ACPI table field value
where an entire table is in scope.

Some ACPI structures require that a field value is unique across all
structures of the same type. For example, the 'ACPI Processor UID'
field in GIC CPU Interface (GICC) structures. Often, the number of
these structures can only be determined at runtime.

The new methods use linked lists to validate either that all
entries in the linked list are unique, or to validate that
all references in the list are valid.

If the acpiview consistency checks are enabled, for every two duplicate
values an error is reported to the user and every invalid or dead
reference is reported to the user.

Change-Id: I6bed7f17670d4fd8813e607753fe0a6c37460891
Signed-off-by: Krzysztof Koch <krzysztof.koch@arm.com>
Signed-off-by: Tomas Pilar <Tomas.Pilar@arm.com>