ShellPkg: acpiview: Validate ID mappings in IORT
Validate that Output References in ID mappings are valid for every node
in the IO Remapping Table (IORT).

The IORT specifications restricts the possible combinations of
cross-references between IORT nodes. This patch adds checks that ensure
that ID mappings between any two nodes are conforming to the
specification given the nodes' type values.

In addition to this, self-references (loops) and references to IORT
nodes which do not exist are detected and reported to the user.

This feature can be switched off using the -q command line flag.

References:
- IO Remapping Table, Platform Design Document, Issue D, March 2018

Change-Id: I82ba5274f9e93205dc835a9f278072f4ff329baa
Signed-off-by: Krzysztof Koch <krzysztof.koch@arm.com>
Signed-off-by: Tomas Pilar <Tomas.Pilar@arm.com>