ShellPkg/AcpiView: Add application wrapper
Simple application wrapper that invokes the shell
command wrapper for acpiview.

This allows the AcpiView functionality to be used
on platforms with older specifications of the UEFI
shell or where the 'acpiview' command is not built in
due to platform build configuration.

Furthermore, this app can be integrated into more comprehensive
testing frameworks as a single component of a more thorough
specification compliance validation strategy.

Change-Id: I564bac8ca8a008aecec476e2b85f590fdee89ed8
Signed-off-by: Tomas Pilar <tomas.pilar@arm.com>