ShellPkg: acpiview: Make GTDT parsing logic data driven
Replace the switch statement in the main parser loop with a table-driven
approach. Use the ParseAcpiStruct () method to resolve how each
Platform Timer Structure given should be parsed.

Enumerate all structures found in the Generic Timer Description Table
(GTDT) on a per-type basis. Print the offset from the start of the table
for each structure.

Consolidate all metadata about each Platform Timer Structure.
Define an array of ACPI_STRUCT_INFO structures to store the name,
instance count, architecture support and handling information. Use this
array to construct the ACPI_STRUCT_DATABASE for GTDT.

Count the number of instances of each Platform Timer Structure type.
Optionally report these counts after GTDT table parsing is finished.

Modify DumpGTBlock () funtion signature so that it matches that of
ACPI_STRUCT_PARSER_FUNC. This way, the function can be used in the
ParseAcpiStruct () call.

Remove the definition of the DumpWatchdogTimer (). Its only purpose was
to call ParseAcpi () and now this process is streamlined.

References:
- ACPI 6.3 Specification - January 2019, Section 5.2.24

Change-Id: I82c0bca26f8c213283021e46851347a67aae75a7
Signed-off-by: Krzysztof Koch <krzysztof.koch@arm.com>
Signed-off-by: Tomas Pilar <Tomas.Pilar@arm.com>