MdeModulePkg/UefiLib: Add PrintSerial()
Add a Print()-equivaltent method that prints a formatted unicode
string to the first serial device found in the platform that provides
a SerialIo protocol.

This functionality is useful as an information stream supplementary
to the consoles defined in the system table, for development and
debugging when the consoles are not available in the platform or
not yet operational in early stages of the boot process.

Change-Id: I8246be533f5a1ee600707c2fd16c119adc3f8acd