DynamicTablesPkg: Add ConfigurationManagerDxe
Add simple ConfigurationManagerDxe reference implementation
of a driver providing a ConfigurationManagerProtocol. The
driver uses a doubly-linked list to store configuration
objects and provides GetObject and SetObject methods for the
protocol. The platform database in the form of the linked list
is not provided in the installed protocol, the field is left
NULL instead.

Change-Id: I961e44a10b75a02ec35af2bbdd9d2579cd767690