DynamicTablesPkg: Update ConfigurationManagerProtocol
The Configuration Manager Protocol is updated to include
a FreeObject function that must be used by the callers to
GetObject to clean up any dynamic allocations and other resources
reserved by the Configuration Manager in the process of fulfilling the
request in GetObject.

A NULL inline static FreeObject function is provided for the ease
of transition for v1.0 managers.

Change-Id: I8ebe3c2b72877e57b2634253c5d4a7c5a8b41141