DynamicTablesPkg: Fold Namespaces into CmObjectId Enums
The enums for CmObjectIds defined in Dynamic Tables Framework
that are used to identify types of configuration manager objects
already include their namespaces in the symbols for enum values.

Namespace enum values are shifted up by NAMESPACE_BIT_SHIFT and
the enums tables for CmObjectIds are bitwise-ORed with their
corresponding namespace.

This means we no longer need to use macros to compose and decompose
CmObjectIds. The macros are adjusted so that they result in NOP.

Change-Id: I3f5cc3b8ae2fc02400e612580fc6b74da40f9e61