ShellPkg/AcpiView: Detect cycles in PPTT
The Processor Hierarchy Node structures contain a reference
to a parent node. The Cache Type structures contain a reference
to the next level of cache.

Both of these fields can be misconfigured as to create a loop
in the chain of references. Such loops can cause OS loaders to
halt, hang or otherwise behave aberrantly.

This patch adds a simple loop detection by indexing all structures
in the PPTT table and then ensuring that any chain of references
is smaller than the total number of PPTT structures found.

Change-Id: Ie7eff4c6e287e24137d964e28b3fb762b723b0fa
Signed-off-by: Tomas Pilar <Tomas.Pilar@arm.com>